Support DATABASES env variable in database setup script

+++ setup-test-db.js
@@ -1,40 +1,56 @@
 const knex = require('knex');
 
-const postgres = knex({
-  client: 'postgres',
+// DATABASES environment variable can contain a comma separated list
+// of databases to setup.
+const DATABASES = (process.env.DATABASES && process.env.DATABASES.split(',')) || [];
 
-  connection: {
-    user: 'postgres',
-    host: 'localhost',
-    database: 'postgres'
-  }
-});
+const knexes = [];
+const commands = [];
 
-const mysql = knex({
-  client: 'mysql',
+if (DATABASES.length === 0 || DATABASES.includes('postgres')) {
+  const postgres = knex({
+    client: 'postgres',
 
-  connection: {
-    user: 'root',
-    host: 'localhost'
-  }
-});
+    connection: {
+      user: 'postgres',
+      host: 'localhost',
+      database: 'postgres',
+    },
+  });
 
-[
-  postgres.raw('DROP DATABASE IF EXISTS objection_test'),
-  postgres.raw('DROP USER IF EXISTS objection'),
-  postgres.raw('CREATE USER objection SUPERUSER'),
-  postgres.raw('CREATE DATABASE objection_test'),
+  knexes.push(postgres);
+  commands.push(
+    postgres.raw('DROP DATABASE IF EXISTS objection_test'),
+    postgres.raw('DROP USER IF EXISTS objection'),
+    postgres.raw('CREATE USER objection SUPERUSER'),
+    postgres.raw('CREATE DATABASE objection_test')
+  );
+}
 
-  mysql.raw('DROP DATABASE IF EXISTS objection_test'),
-  mysql.raw('DROP USER IF EXISTS objection'),
-  mysql.raw('CREATE USER objection'),
-  mysql.raw('GRANT ALL PRIVILEGES ON *.* TO objection'),
-  mysql.raw('CREATE DATABASE objection_test')
-].reduce((promise, query) => {
-  return promise.then(() => query);
-}, Promise.resolve()).then(() => {
-  return Promise.all([
-    postgres.destroy(),
-    mysql.destroy()
-  ]);
-});
+if (DATABASES.length === 0 || DATABASES.includes('mysql')) {
+  const mysql = knex({
+    client: 'mysql',
+
+    connection: {
+      user: 'root',
+      host: 'localhost',
+    },
+  });
+
+  knexes.push(mysql);
+  commands.push(
+    mysql.raw('DROP DATABASE IF EXISTS objection_test'),
+    mysql.raw('DROP USER IF EXISTS objection'),
+    mysql.raw('CREATE USER objection'),
+    mysql.raw('GRANT ALL PRIVILEGES ON *.* TO objection'),
+    mysql.raw('CREATE DATABASE objection_test')
+  );
+}
+
+commands
+  .reduce((promise, query) => {
+    return promise.then(() => query);
+  }, Promise.resolve())
+  .then(() => {
+    return Promise.all(knexes.map((it) => it.destroy()));
+  });
